Refactor per instructions.
Here's the instructions:

This solution contains three classes used by an imaginary bicycle distributor 
to produce order receipts and some unit tests to prove that everything works.

Pretend this code is part of a larger software system and you are given 
responsibility for it. Assume that there will be at least one type of 
change coming your way regularly: 
	more bikes at new prices, 
	different discount codes and percentages, 
	and/or additional receipt formats.

Refactor the code so that it can survive an onslaught of the changes you've 
chosen, you're confident it works, and you're comfortable the next engineer 
will easily understand how to work on it.